+++ ls_Glass/core/changelog.lua
@@ -6,13 +6,8 @@ local _G = getfenv(0)
 
 -- Mine
 E.CHANGELOG = [[
-- Added 10.0.2 support.
-- Increased the max fade out delay to 120s.
-- Changed the initial anchor of the edit box. It now can be positioned inside the chat frame.
-- Fixed an issue where profession links would open a profession window on mouse over. Thanks
-  truckcarr11@GitHub.
-- Updated French translation. Translated by Braincell1980@Curse.
-- Updated German translation. Translated by OHerendirO@Curse.
-- Updated Korean translation. Translated by netaras@Curse.
-- Updated Traditional Chinese translation. Translated by RainbowUI@Curse.
+- Added 10.0.5 support.
+- Added optional buttons for scrolling up and down. It's an accessibility feature meant for users
+  who can't use the mouse wheel. Holding the Ctrl key will slow the scrolling rate by two times.
+- Added an option to fade in chat messages on mouseover.
 ]]